remove CreateLogic
